--- app/src/main/java/sahak/sahakyan/dynamicdeals/MainActivity.kt
@@ -1,26 +1,0 @@
-package sahak.sahakyan.dynamicdeals
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import sahak.sahakyan.dynamicdeals.ui.theme.DynamicDealsTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            DynamicDealsTheme {
-
-            }
-        }
-    }
-}
