+++ app/src/main/java/sahak/sahakyan/dynamicdeals/presentation/ui/MainActivity.kt
@@ -1,0 +1,19 @@
+package sahak.sahakyan.dynamicdeals.presentation.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import sahak.sahakyan.dynamicdeals.presentation.ui.theme.DynamicDealsTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            DynamicDealsTheme {
+
+            }
+        }
+    }
+}
